refactor(std): add std/_util/os.ts module (#8370)
This commit adds "std/_util/os.ts" module which contains
common os detection logic and is browser compatible.
