+++ std/path/_constants.ts
@@ -46,17 +46,3 @@ export const CHAR_EQUAL = 61; /* = */
 // Digits
 export const CHAR_0 = 48; /* 0 */
 export const CHAR_9 = 57; /* 9 */
-
-let NATIVE_OS: typeof Deno.build.os = "linux";
-// deno-lint-ignore no-explicit-any
-const navigator = (globalThis as any).navigator;
-if (globalThis.Deno != null) {
-  NATIVE_OS = Deno.build.os;
-} else if (navigator?.appVersion?.includes?.("Win") ?? false) {
-  NATIVE_OS = "windows";
-}
-// TODO(nayeemrmn): Improve OS detection in browsers beyond Windows.
-
-export const isWindows = NATIVE_OS == "windows";
-
-export { NATIVE_OS };
