+++ std/path/glob.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 // This module is browser compatible.
 
-import { NATIVE_OS } from "./_constants.ts";
+import { osType } from "../_util/os.ts";
 import { join, normalize } from "./mod.ts";
 import { SEP, SEP_PATTERN } from "./separator.ts";
 
@@ -80,7 +80,7 @@ const rangeEscapeChars = ["-", "\\", "]"];
  *   the group occurs not nested at the end of the segment. */
 export function globToRegExp(
   glob: string,
-  { extended = true, globstar: globstarOption = true, os = NATIVE_OS }:
+  { extended = true, globstar: globstarOption = true, os = osType }:
     GlobToRegExpOptions = {},
 ): RegExp {
   if (glob == "") {
