+++ std/path/mod.ts
@@ -2,7 +2,7 @@
 // Ported mostly from https://github.com/browserify/path-browserify/
 /** This module is browser compatible. */
 
-import { isWindows } from "./_constants.ts";
+import { isWindows } from "../_util/os.ts";
 import * as _win32 from "./win32.ts";
 import * as _posix from "./posix.ts";
