+++ std/path/separator.ts
@@ -1,7 +1,7 @@
 // Copyright 2018-2020 the Deno authors. All rights reserved. MIT license.
 /** This module is browser compatible. */
 
-import { isWindows } from "./_constants.ts";
+import { isWindows } from "../_util/os.ts";
 
 export const SEP = isWindows ? "\\" : "/";
 export const SEP_PATTERN = isWindows ? /[\\/]+/ : /\/+/;
